Renames DateTime to Timestamp. Not sure why I felt so motivated.
