+++ data/src/Format.ts
@@ -1,6 +1,6 @@
 export type { Format } from "./Format/generate";
 
-export * as DateTime from "./Format/DateTime";
 export * as Noid from "./Format/Noid";
 export * as Slug from "./Format/Slug";
+export * as Timestamp from "./Format/Timestamp";
 export * as UriReference from "./Format/UriReference";
