+++ data/src/Util/ProcessUpdate.ts
@@ -1,13 +1,13 @@
 import { JSONSchemaType } from "ajv";
 import Schema from "../Schema";
-import { DateTime, schema as dateTimeSchema } from "../Format/DateTime";
+import { Timestamp, schema as timestampSchema } from "../Format/Timestamp";
 
 /**
  * An object that describes a request for and the output of an automated process that is applied to the parent access object.
  */
 export interface ProcessUpdate {
-  requestDate: DateTime;
-  processDate?: DateTime;
+  requestDate: Timestamp;
+  processDate?: Timestamp;
   succeeded?: boolean;
   message?: string;
 }
@@ -17,8 +17,8 @@ export const schema = new Schema({
   title: "Process update",
   type: "object",
   properties: {
-    requestDate: dateTimeSchema.inline,
-    processDate: { ...dateTimeSchema.inline, nullable: true },
+    requestDate: timestampSchema.inline,
+    processDate: { ...timestampSchema.inline, nullable: true },
     succeeded: { type: "boolean", nullable: true },
     message: { type: "string", nullable: true },
   },
